moved pomodoro and fixed up various smaller style issues
