+++ client/src/components/navbar.tsx
@@ -8,6 +8,7 @@ import { logout } from '../api/loginRegisterOperation';
 import Button from 'react-bootstrap/Button';
 // import bootstrap css
 import  'bootstrap/dist/css/bootstrap.min.css';
+import logo from '../assets/logo.png';
 
 
 /**
@@ -29,7 +30,16 @@ export function NavigationPanel() {
     return (
         <Navbar expand="lg" className="bg-body-tertiary">
             <Container fluid>
-                <Navbar.Brand href="todo">Brand name</Navbar.Brand>
+                <Navbar.Brand href="todo">
+                    <img
+                        src={logo}
+                        alt="Logo"
+                        width="40" // Adjust size as needed
+                        height="40"
+                        className="d-inline-block align-top"
+                    />
+                    {' '}Study app
+                </Navbar.Brand>
                 <Navbar.Toggle aria-controls="navbarScroll" />
                 <Navbar.Collapse id="navbarScroll">
                     {(function() {
@@ -49,35 +59,34 @@ export function NavigationPanel() {
                                 <Nav.Link href="/pomodoro" >
                                     Pomodoro
                                 </Nav.Link>
-        
-        
-        
-        
                             </Nav>
                             );
                         }
                         else {
                             return (
                                 <Nav
-                                className="me-auto my-2 my-lg-0"
-                                style={{ maxHeight: '100px' }}
-                                navbarScroll
-                            >
-
-                            </Nav>
+                                    className="me-auto my-2 my-lg-0"
+                                    style={{ maxHeight: '100px' }}
+                                    navbarScroll
+                                >
+                                </Nav>
                             );
                         }
                     })()}
 
                 </Navbar.Collapse>
                 <Navbar.Collapse className="justify-content-end">
-                    <Navbar.Text className="logged-in-text">
-                    Logged in as: {localStorage.getItem('username')}
-                    </Navbar.Text>
                     {(()=>{if (localStorage.getItem('username')) {
-                        return (<Button variant="outline-success" onClick={logout}>
-                            Logout
-                        </Button>)
+                        return (
+                        <>
+                            <Navbar.Text className="logged-in-text">
+                                Logged in as: {localStorage.getItem('username')}
+                            </Navbar.Text>
+                            <Button variant="outline-success" onClick={logout}>
+                                Logout
+                            </Button>
+                        </>
+                        )
                     }
                     else{
                         return (<Button variant="outline-success" href="/">
