--- client/src/pomodoro.css
@@ -1,51 +1,0 @@
-/* Base styles */
-body {
-    font-family: Arial, sans-serif;
-    background-color: #000000;
-    margin: 0;
-    padding: 0;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-}
-
-/* Timer container */
-.timer-container {
-    background-color: #4d4d4d;
-    border-radius: 10px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    padding: 20px;
-    text-align: center;
-    width: 300px;
-}
-
-/* Timer display */
-.timer-display {
-    font-size: 48px;
-    font-weight: bold;
-    margin-bottom: 20px;
-    
-}
-
-/* Buttons */
-button {
-    background-color: #ff6b6b;
-    border: none;
-    border-radius: 5px;
-    color: #fff;
-    cursor: pointer;
-    font-size: 16px;
-    margin: 5px;
-    padding: 10px 20px;
-    transition: background-color 0.3s ease;
-}
-
-button:hover {
-    background-color: #ff4c4c;
-}
-
-button:disabled {
-    background-color: #ccc;
-    cursor: not-allowed;
-}
